+++ GoFr Server/configs/.env
@@ -1,0 +1,1 @@
+HTTP_PORT=9000
