--- GoFr Server/twitter_tokens.json
@@ -1,1 +1,0 @@
-{"access_token": "1860400181059125248-yCt2VWZHa9SLZf8eBbASyiZb4VMwe4", "access_token_secret": "vDdx0wKyV20xVw6U7QuIRObfdzDVIAilhbE4loKSMGCo3"}
